Add files via upload
Handball is a team sport played with a ball, where two teams of seven players (including a goalkeeper) compete to score the most goals during the game. The gameplay combines rapid attacks and strategic defense. We are developing a support application to enhance understanding and practice of the game, assisting teams with information and protection for effective play.